Check if certificate was requested from current node and skip creation of chains if not.
This is necessary because otherwise all nodes would try to create the chain.
However this would fail on every node other than the node that requests the certificate because the private key or dh file are not present.

fixes #31

Not sure if `$domain` is the correct variable to check for in `$acme_csrs_array` due to wildcard replacement.